rename and correct mistakes in selectYear
