+++ index.js
@@ -1,0 +1,11 @@
+import { windowForUserScript } from './modules/selectYear.js';
+import { defaultMD } from './modules/autoMD.js';
+import { addBattons } from './modules/addBattons.js';
+import { addTamplaties } from './modules/addTamplaties.js';
+import { usersSettings } from './modules/usersSettings.js'
+import { checkReload } from './modules/checkReload.js'
+import { statisticsInterface } from './modules/statisticsInterface.js'
+import { addStaticOnLocalStorage } from './modules/addStaticOnLocalStorage.js'
+
+
+export { windowForUserScript, defaultMD, addBattons, addTamplaties, addStaticOnLocalStorage, usersSettings, checkReload, statisticsInterface }
