+++ kad.js
@@ -1,15 +1,25 @@
 ﻿
 "use strict";
 
-import { windowForUserScript } from './modules/alert.js';
-import { defaultMD } from './modules/autoMD.js';
-import { addBattons } from './modules/addBattons.js';
-import { addTamplaties } from './modules/addTamplaties.js';
-import { usersSettings } from './modules/usersSettings.js'
-import { checkReload } from './modules/checkReload.js'
-import { statisticsInterface } from './modules/statisticsInterface.js'
-import { addStaticOnLocalStorage } from './modules/addStaticOnLocalStorage.js'
+// import { windowForUserScript } from './modules/alert.js';
+// import { defaultMD } from './modules/autoMD.js';
+// import { addBattons } from './modules/addBattons.js';
+// import { addTamplaties } from './modules/addTamplaties.js';
+// import { usersSettings } from './modules/usersSettings.js'
+// import { checkReload } from './modules/checkReload.js'
+// import { statisticsInterface } from './modules/statisticsInterface.js'
+// import { addStaticOnLocalStorage } from './modules/addStaticOnLocalStorage.js'
 
+import {
+  windowForUserScript,
+  defaultMD,
+  addBattons,
+  addTamplaties,
+  addStaticOnLocalStorage,
+  usersSettings,
+  checkReload,
+  statisticsInterface
+} from './index.js'
 
 if (document.location.href === `https://kad.arbitr.ru/`) { windowForUserScript() }
